+++ lib/riverpie.dart
@@ -1,5 +1,7 @@
 export 'package:riverpie/src/mixin.dart';
 export 'package:riverpie/src/notifier.dart';
+export 'package:riverpie/src/observer/event.dart';
+export 'package:riverpie/src/observer/observer.dart';
 export 'package:riverpie/src/provider/provider.dart';
 export 'package:riverpie/src/provider/custom/future_provider.dart';
 export 'package:riverpie/src/provider/custom/state_provider.dart';