+++ lib/src/widget/consumer.dart
@@ -4,20 +4,27 @@ import 'package:riverpie/src/ref.dart';
 
 /// A [Consumer] can be used anywhere in the widget tree.
 /// This is useful if you want to use a provider within a [StatelessWidget].
+///
+/// Add a [debugLabel] or [debugParent] for better logging.
 class Consumer extends StatelessWidget {
+  final String debugLabel;
   final Widget Function(
     BuildContext context,
     WatchableRef ref,
   ) builder;
 
-  const Consumer({
+  Consumer({
     super.key,
+    String? debugLabel,
+    Widget? debugParent,
     required this.builder,
-  });
+  }) : debugLabel =
+            debugLabel ?? debugParent?.runtimeType.toString() ?? 'Consumer';
 
   @override
   Widget build(BuildContext context) {
     return ExpensiveConsumer(
+      debugLabel: debugLabel,
       builder: (context, ref, _) {
         return builder(context, ref);
       },
@@ -28,7 +35,10 @@ class Consumer extends StatelessWidget {
 /// Similar to [Consumer] but with a [child]
 /// that is not rebuilt when the provider changes.
 /// This is useful if the [child] is expensive to build.
+///
+/// Add a [debugLabel] or [debugParent] for better logging.
 class ExpensiveConsumer extends StatefulWidget {
+  final String debugLabel;
   final Widget? child;
   final Widget Function(
     BuildContext context,
@@ -36,11 +46,15 @@ class ExpensiveConsumer extends StatefulWidget {
     Widget? child,
   ) builder;
 
-  const ExpensiveConsumer({
+  ExpensiveConsumer({
     super.key,
+    String? debugLabel,
+    Widget? debugParent,
     this.child,
     required this.builder,
-  });
+  }) : debugLabel = debugLabel ??
+            debugParent?.runtimeType.toString() ??
+            'ExpensiveConsumer';
 
   @override
   State<ExpensiveConsumer> createState() => _ExpensiveConsumerState();